Revert "Fix ESLint linebreak-style errors by preserving line endings in LSP communication (#38773)"
This reverts commit 435eab68968fa11f8a476c78845b13df45498f25.
