+++ crates/text/Cargo.toml
@@ -23,6 +23,7 @@ log.workspace = true
 parking_lot.workspace = true
 postage.workspace = true
 rand = { workspace = true, optional = true }
+regex.workspace = true
 rope.workspace = true
 smallvec.workspace = true
 sum_tree.workspace = true
